support BigInt schema type (64-bit int)
Implemented this change so that the page table can have offsets greater
than maxInt32.

All SQL integer values are stored as int64 in application code. This
allows us to handle both 32-bit and 64-bit values without adding
complicated logic. We are explicit about the type, as opposed to
simply using the int type, in order to support 64-bit integers on
32-bit systems.

When we serialize 32-bit integers before writing out to disk, we
ensure that the int64 value is within 32-bit range, and the value
is written out as a 32-bit integer, which saves space.